@@ -1,11 +1,8 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-   content: [
-      "./index.html",
-      "./src/**/*.{ts,tsx}"
-   ],
+export default {
+   content: ["./index.html", "./src/**/*.{ts,tsx}"],
    theme: {
       extend: {},
    },
    plugins: [],
-}
+};
