@@ -1,7 +1,36 @@
+import { useState } from "react";
+import reactLogo from "./assets/react.svg";
+import viteLogo from "/vite.svg";
 import "./App.css";
 
 function App() {
-   return <h1 className="text-3xl font-bold underline">test</h1>;
+   const [count, setCount] = useState(0);
+
+   return (
+      <>
+         <div>
+            <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
+               <img src={viteLogo} className="logo" alt="Vite logo" />
+            </a>
+            <a href="https://react.dev" target="_blank" rel="noreferrer">
+               <img src={reactLogo} className="logo react" alt="React logo" />
+            </a>
+         </div>
+         <h1 className="text-3xl font-bold underline">Vite + React</h1>
+         <div className="card">
+            {/* biome-ignore lint/a11y/useButtonType: <explanation> */}
+            <button onClick={() => setCount((count) => count + 1)}>
+               count is {count}
+            </button>
+            <p>
+               Edit <code>src/App.tsx</code> and save to test HMR
+            </p>
+         </div>
+         <p className="read-the-docs">
+            Click on the Vite and React logos to learn more
+         </p>
+      </>
+   );
 }
 
 export default App;
